Rename contants.js into constants.js
